Added csv files for organisms.
- Changed how cell width and height are calculated. Float types now
used.
- Added simple debugging function
- Button selection clears when a key is pressed.

+++ RadioButton.pde
@@ -5,26 +5,26 @@ class RadioButton
   color baseColor = color (239, 239, 239);
   color checkColor = color (60, 60, 60);
   color textColor = color(0);
-  int x, y;
+  float x, y;
   boolean checked;
   String text;
-  
+
   RadioButton(String txt)
   {
     this(-1, -1, txt);
   }
-  RadioButton(int xLoc, int yLoc, String txt)
+  RadioButton(float xLoc, float yLoc, String txt)
   {
     this(xLoc, yLoc, txt, false);
   }
-  RadioButton(int xLoc, int yLoc, String txt, boolean defChk)
+  RadioButton(float xLoc, float yLoc, String txt, boolean defChk)
   {
     text = txt;
     x = xLoc;
     y = yLoc;
     checked = defChk;
   }
-  
+
   void displayGradient()
   {
     noStroke();
@@ -40,28 +40,26 @@ class RadioButton
       ellipse(x, y, r, r);
     }
   }
-  
+
   void display()
   {
     displayGradient();
-    if(checked)
+    if (checked)
     {
       noStroke();
       fill(checkColor);
       ellipse(x, y, size/2, size/2);
     }  
-    
+
     fill(textColor);
     text(text, x + size + 5, y + 5);
   }
-  
+
   boolean isInButton(int xLoc, int yLoc)
   {
     float disX = x - xLoc;
     float disY = y - yLoc;
     return sqrt(sq(disX) + sq(disY)) < (size / 2);
   }
-
 }
-
